requirements.txt
# VittaSaathi v3.0 - WhatsApp Financial Advisor
# Core Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.26.3

# NLP & ML
# prophet==1.1.5  # Uncomment for income prediction (requires cmdstan)

# OCR & Document Processing
pytesseract==0.3.10
Pillow==10.2.0
PyPDF2==3.0.1

# Voice & TTS
gtts==2.5.0

# Communication
twilio==8.13.0

# Database (using JSON for simplicity, upgrade to SQLite/PostgreSQL as needed)
# For production, uncomment:
# sqlalchemy==2.0.25
# alembic==1.13.1

# Scheduling
apscheduler==3.10.4

# Environment
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2

# PDF Generation
reportlab==4.0.8

# Excel Export
openpyxl==3.1.2

# Encryption (for secure backups)
cryptography==41.0.7

# Cloud Backup (optional)
boto3==1.34.29
# google-cloud-storage==2.14.0  # Uncomment for GCS support

# Two-Factor Authentication
pyotp==2.9.0
qrcode[pil]==7.4.2

# OpenAI (for voice transcription and smart NLP)
openai==1.12.0
